+++ actix-files/CHANGES.md
@@ -1,6 +1,14 @@
 # Changes
 
-## [0.1.0] - 2019-05-xx
+## [0.1.2] - 2019-06-06
+
+* Fix ring dependency from actix-web default features for #741.
+
+## [0.1.1] - 2019-06-01
+
+* Static files are incorrectly served as both chunked and with length #812
+
+## [0.1.0] - 2019-05-25
 
 * NamedFile last-modified check always fails due to nano-seconds
   in file modified date #820