+++ actix-multipart/CHANGES.md
@@ -1,12 +1,24 @@
 # Changes
 
+## [0.1.3] - 2019-06-06
+
+* Fix ring dependency from actix-web default features for #741.
+
+## [0.1.2] - 2019-06-02
+
+* Fix boundary parsing #876
+
+## [0.1.1] - 2019-05-25
+
+* Fix disconnect handling #834
+
 ## [0.1.0] - 2019-05-18
 
 * Release
 
 ## [0.1.0-beta.4] - 2019-05-12
 
-* Handle cancellation of uploads #834 #736
+* Handle cancellation of uploads #736
 
 * Upgrade to actix-web 1.0.0-beta.4