+++ actix-web-codegen/CHANGES.md
@@ -1,5 +1,13 @@
 # Changes
 
+## [0.1.2] - 2019-06-04
+
+* Add macros for head, options, trace, connect and patch http methods
+
+## [0.1.1] - 2019-06-01
+
+* Add syn "extra-traits" feature
+
 ## [0.1.0] - 2019-05-18
 
 * Release